requirements.txt
anaconda-client @ file:///C:/ci/anaconda-client_1624480273070/work
anaconda-navigator==2.0.3
appdata==2.1.2
argon2-cffi @ file:///C:/ci/argon2-cffi_1613037959010/work
async-generator @ file:///home/ktietz/src/ci/async_generator_1611927993394/work
atomicwrites==1.4.0
attrs @ file:///tmp/build/80754af9/attrs_1620827162558/work
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
backports.functools-lru-cache @ file:///tmp/build/80754af9/backports.functools_lru_cache_1618170165463/work
backports.tempfile @ file:///home/linux1/recipes/ci/backports.tempfile_1610991236607/work
backports.weakref==1.0.post1
beautifulsoup4 @ file:///home/linux1/recipes/ci/beautifulsoup4_1610988766420/work
bleach @ file:///tmp/build/80754af9/bleach_1628110601003/work
brotlipy==0.7.0
certifi==2021.5.30
cffi @ file:///C:/ci/cffi_1625831749120/work
chardet==3.0.4
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.0.3
clyent==1.2.2
colorama @ file:///tmp/build/80754af9/colorama_1607707115595/work
conda==4.10.3
conda-build==3.21.4
conda-content-trust @ file:///tmp/build/80754af9/conda-content-trust_1617045594566/work
conda-package-handling @ file:///C:/ci/conda-package-handling_1618262320430/work
conda-repo-cli @ file:///tmp/build/80754af9/conda-repo-cli_1620168426516/work
conda-token @ file:///tmp/build/80754af9/conda-token_1620076980546/work
conda-verify==3.4.2
cryptography @ file:///C:/ci/cryptography_1616769344312/work
debugpy @ file:///C:/ci/debugpy_1629222854708/work
decorator @ file:///tmp/build/80754af9/decorator_1621259047763/work
defusedxml @ file:///tmp/build/80754af9/defusedxml_1615228127516/work
Django==2.2.16
django-debug-toolbar==2.2
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
entrypoints==0.3
fake-useragent==0.1.11
Faker==12.0.1
filelock @ file:///home/linux1/recipes/ci/filelock_1610993975404/work
future==0.18.2
glob2 @ file:///home/linux1/recipes/ci/glob2_1610991677669/work
idna==2.8
importlib-metadata @ file:///C:/ci/importlib-metadata_1631053185730/work
iniconfig==1.1.1
ipykernel @ file:///C:/ci/ipykernel_1629995240690/work/dist/ipykernel-6.2.0-py3-none-any.whl
ipython @ file:///C:/ci/ipython_1628244008373/work
ipython-genutils @ file:///tmp/build/80754af9/ipython_genutils_1606773439826/work
jedi @ file:///C:/ci/jedi_1611315808330/work
Jinja2 @ file:///tmp/build/80754af9/jinja2_1612213139570/work
jsonschema @ file:///Users/ktietz/demo/mc3/conda-bld/jsonschema_1630511932244/work
jupyter-client @ file:///tmp/build/80754af9/jupyter_client_1630690655192/work
jupyter-core @ file:///C:/ci/jupyter_core_1612213356021/work
jupyterlab-pygments @ file:///tmp/build/80754af9/jupyterlab_pygments_1601490720602/work
libarchive-c @ file:///tmp/build/80754af9/python-libarchive-c_1617780486945/work
LiveWires==2.1
MarkupSafe @ file:///C:/ci/markupsafe_1621528314575/work
matplotlib-inline @ file:///tmp/build/80754af9/matplotlib-inline_1628242447089/work
menuinst==1.4.16
mistune==0.8.4
mixer==7.1.2
more-itertools==8.13.0
navigator-updater==0.2.1
nbclient @ file:///tmp/build/80754af9/nbclient_1614364831625/work
nbconvert @ file:///C:/ci/nbconvert_1624479163777/work
nbformat @ file:///tmp/build/80754af9/nbformat_1617383369282/work
nest-asyncio @ file:///tmp/build/80754af9/nest-asyncio_1613680548246/work
notebook @ file:///C:/ci/notebook_1629205926310/work
olefile @ file:///Users/ktietz/demo/mc3/conda-bld/olefile_1629805411829/work
packaging @ file:///tmp/build/80754af9/packaging_1625611678980/work
pandocfilters @ file:///C:/ci/pandocfilters_1605102497129/work
parso @ file:///tmp/build/80754af9/parso_1617223946239/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow==8.3.1
pkginfo==1.7.1
pluggy==0.13.1
prometheus-client @ file:///tmp/build/80754af9/prometheus_client_1623189609245/work
prompt-toolkit @ file:///Users/ktietz/demo/mc3/conda-bld/prompt-toolkit_1629146841569/work
psutil @ file:///C:/ci/psutil_1612298324802/work
py==1.11.0
pycosat==0.6.3
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
pygame==2.1.2
Pygments @ file:///tmp/build/80754af9/pygments_1629234116488/work
PyJWT==2.1.0
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1608057966937/work
pyparsing @ file:///home/linux1/recipes/ci/pyparsing_1610983426697/work
pyrsistent @ file:///C:/ci/pyrsistent_1600141795814/work
PySocks @ file:///C:/ci/pysocks_1605287845585/work
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
pytils==0.3
pytz==2019.3
pywin32==228
pywinpty==0.5.7
PyYAML==5.4.1
pyzmq @ file:///C:/ci/pyzmq_1628276144921/work
QtPy @ file:///tmp/build/80754af9/qtpy_1629397026935/work
requests==2.26.0
ruamel-yaml-conda @ file:///C:/ci/ruamel_yaml_1616016967756/work
Send2Trash @ file:///tmp/build/80754af9/send2trash_1607525499227/work
sip==4.19.13
six==1.14.0
sorl-thumbnail==12.7.0
soupsieve @ file:///tmp/build/80754af9/soupsieve_1616183228191/work
sqlparse==0.4.2
SuperWires==0.9.6
terminado==0.9.4
testpath @ file:///tmp/build/80754af9/testpath_1624638946665/work
to-requirements.txt==1.1.10
toml==0.10.2
tornado @ file:///C:/ci/tornado_1606942392901/work
tqdm @ file:///tmp/build/80754af9/tqdm_1629302309755/work
traitlets @ file:///home/ktietz/src/ci/traitlets_1611929699868/work
urllib3==1.25.11
wcwidth @ file:///Users/ktietz/demo/mc3/conda-bld/wcwidth_1629357192024/work
webencodings==0.5.1
win-inet-pton @ file:///C:/ci/win_inet_pton_1605306167264/work
wincertstore==0.2
xmltodict @ file:///Users/ktietz/demo/mc3/conda-bld/xmltodict_1629301980723/work
zipp @ file:///tmp/build/80754af9/zipp_1625570634446/work
